feat: 메인 페이지를 클라이언트 사이드 렌더링으로 변경
- 서버 사이드 렌더링(SSR)에서 클라이언트 사이드 렌더링(CSR)으로 전환
- 새로운 home-client.tsx 컴포넌트 생성하여 동적 데이터 로딩 구현
- API 엔드포인트를 통해 최신 포스트 3개를 런타임에 가져옴
- 로딩 상태, 에러 처리, 재시도 기능 추가
- AbortController를 사용한 요청 취소 처리 구현

이제 새 포스트가 추가되어도 재빌드 없이 즉시 반영됩니다.
